Middleware
Added middleware to authorize users

+++ routes.php
@@ -1,16 +1,16 @@
 <?php
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
 
-$router->post('/listings', 'ListingController@store');
-$router->delete('/listings/{id}', 'ListingController@destroy');
-$router->put('/listings/update/{id}', 'ListingController@update');
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
+$router->put('/listings/update/{id}', 'ListingController@update', ['auth']);
 
-$router->get('/auth/register', 'UserController@register');
+$router->get('/auth/register', 'UserController@register', ['guest']);
 $router->post('/auth/register', 'UserController@registerUser');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/login', 'UserController@login', ['guest']);
 $router->post('/auth/login', 'UserController@authenticate');
-$router->post('/auth/logout', 'UserController@logout');
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
